Changes: 1) Changed project name from "accounts3" to "AccountsB52z" | "accountsb52z" 2) Updated the package throughout the files. 3) Cleared most of the warnings.

+++ lib/screens/home/common_variables_homepage.dart
@@ -2,10 +2,10 @@ double balanceFundTotalPulledFromDB = 0;
 double totalFundRecievedPulledFromDB = 0;
 double totalFundExpectedCalculatedLocal = 0;
 
-int switchCaseRetrievedValueMonthly = 0;
+double switchCaseRetrievedValueMonthly = 0;
 double switchCaseRetrievedValueLoan = 0;
 
-int totalMonthlyPendingValueAllMembersPendingAmountCalcFromListMemberWise = 0;
+double totalMonthlyPendingValueAllMembersPendingAmountCalcFromListMemberWise = 0;
 
 List<int> pendingCountListMemberWiseMonthly = [];
 List<double> pendingLoanAmountAllMembersPulledDB = [];
